[Mod] Make SmpLength 32bit even on 64bit architectures. This silences a huge amount of 64bit warnings. We do not support samples that large anyway, and that is not going to change in the forseeable future. Hopefully, nothing will break. [Mod] OpenMPT: Version is now 1.27.00.58
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@8702 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/Snd_defs.h
@@ -36,7 +36,7 @@ typedef uint16 INSTRUMENTINDEX;
 typedef uint8 SEQUENCEINDEX;
 	const SEQUENCEINDEX SEQUENCEINDEX_INVALID = uint8_max;
 
-typedef uintptr_t SmpLength;
+typedef uint32 SmpLength;
 
 
 const SmpLength MAX_SAMPLE_LENGTH	= 0x10000000;	// Sample length in *frames*
